refactor: improve route handler structure
using helper functions from util
