+++ src/app/api/errors.ts
@@ -35,20 +35,3 @@ export class UnauthorizedError extends Error {
     super(message, options);
   }
 }
-
-const errorMap: Record<string, { status: number }> = {
-  DatabaseOperationError: { status: 500 },
-  NotFoundError: { status: 404 },
-  InputParseError: { status: 400 },
-  AuthenticationError: { status: 401 },
-  UnauthenticatedError: { status: 401 },
-  UnauthorizedError: { status: 403 },
-};
-
-export function handleError(error: Error) {
-  const errorType = error.constructor.name;
-  const { status } = errorMap[errorType] || { status: 500 };
-  const message = error.message || "Internal server error";
-
-  return Response.json({ message }, { status });
-}
