+++ src/app/api/utils/index.ts
@@ -1,7 +1,11 @@
 import { validateInput } from "./validation";
-import { validateSession } from "./auth";
+import { getSessionId, validateSession } from "./auth";
+import { errorResponse, successResponse } from "./response";
 
 export default {
   validateInput,
   validateSession,
+  getSessionId,
+  successResponse,
+  errorResponse,
 };
